refactor(ci): preserve latency evidence budgets
