+++ package.json
@@ -86,7 +86,7 @@
     "dev-gate-latency-patrol:classify": "node scripts/require-shifu.mjs dev-gate-latency-patrol:classify && node framework/dev-gate-latency-patrol/classify.mjs",
     "dev-gate-latency-patrol:dogfood-capture": "node scripts/require-shifu.mjs dev-gate-latency-patrol:dogfood-capture && node framework/dev-gate-latency-patrol/dogfood-capture.mjs",
     "ci:cancel-dequeued-merge-group": "node scripts/require-shifu.mjs ci:cancel-dequeued-merge-group && node scripts/cancel-dequeued-merge-group-runs.mjs",
-    "test:gate-latency": "node scripts/require-shifu.mjs test:gate-latency && node --test scripts/affected-native-cache-payload.test.mjs scripts/affected-native-proof.test.mjs scripts/affected-native-semantic-source.test.mjs scripts/cancel-dequeued-merge-group-runs.test.mjs scripts/candidate-timeline-events.test.mjs scripts/dev-delivery-proof-consumer.test.mjs scripts/dev-delivery-warrant-input.test.mjs scripts/dev-gate-latency-artifacts.test.mjs scripts/measure-dev-required-latency.test.mjs scripts/native-execution-under-warrant.test.mjs scripts/queue-admission-lease.test.mjs scripts/run-core-affected-native.test.mjs scripts/write-affected-native-cache-manifests.test.mjs",
+    "test:gate-latency": "node scripts/require-shifu.mjs test:gate-latency && node --test scripts/affected-native-cache-payload.test.mjs scripts/affected-native-proof.test.mjs scripts/affected-native-semantic-source.test.mjs scripts/cancel-dequeued-merge-group-runs.test.mjs scripts/candidate-timeline-events.test.mjs scripts/dev-delivery-proof-consumer.test.mjs scripts/dev-delivery-warrant-input.test.mjs scripts/dev-gate-latency-artifacts.test.mjs scripts/dev-gate-queue-evidence.test.mjs scripts/measure-dev-required-latency.test.mjs scripts/native-execution-under-warrant.test.mjs scripts/queue-admission-lease.test.mjs scripts/run-core-affected-native.test.mjs scripts/write-affected-native-cache-manifests.test.mjs",
     "alpha:publication:tail": "node scripts/require-shifu.mjs alpha:publication:tail && node scripts/alpha-publication-tail-plan.mjs",
     "alpha:release:timeline": "node scripts/require-shifu.mjs alpha:release:timeline && node scripts/alpha-release-timeline.mjs",
     "alpha:release:history": "node scripts/require-shifu.mjs alpha:release:history && node scripts/alpha-release-history.mjs",
