+++ scripts/measure-dev-required-latency.mjs
@@ -12,6 +12,7 @@ import {
   createCandidateTimeline,
   formatCandidateTimelineReport,
 } from '@kungfu-tech/buildchain-alpha/candidate-timeline';
+import * as queueEvidence from '../framework/dev-delivery/dev-gate-queue-evidence.mjs';
 import { latencyBaselineForDevBranch } from '../framework/version-line/version-line-authority.mjs';
 import {
   affectedNativeClassification,
@@ -42,14 +43,18 @@ import {
 import { planAffectedPaths } from './run-core-affected-native.mjs';
 export { requiredMergeQueueWindow };
 export {
+  postMergeAdvisoryEvidence,
+  summarizeMergeQueueDelivery,
+  summarizePostMergeAdvisory,
+} from '../framework/dev-delivery/dev-gate-queue-evidence.mjs';
+export {
   affectedNativeEvidenceBinding,
   deliveryAttemptEvidenceFromMembers,
   finalDevAncestryFromCompare,
   reconstructDeliveryEvidence,
 } from './cancel-dequeued-merge-group-runs.mjs';
 const ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
-const MINIMUM_SAMPLE_COUNT = 20;
-const MINIMUM_NATIVE_SAMPLE_COUNT = 10;
+const [MINIMUM_SAMPLE_COUNT, MINIMUM_NATIVE_SAMPLE_COUNT] = [20, 10];
 const QUEUE_ADMISSION_CONTRACT_PATH = path.join(
   ROOT,
   'docs/qualification/gates/dev-queue-admission.contract.json',
@@ -252,41 +257,6 @@ function projectedJob(job) {
   };
 }
 
-function workflowIdentity(run) {
-  return String(run.workflow_id || run.path || run.name || 'unknown-workflow');
-}
-
-function validationRepeatEvidence(rounds) {
-  const seen = new Set();
-  let repeated = 0;
-  let explained = 0;
-  let unexplained = 0;
-  const explanations = {};
-  for (const round of rounds) {
-    const seenInRound = new Set();
-    for (const run of round.mergeGroupRuns || []) {
-      const identity = run.workflowIdentity || String(run.id);
-      if (seen.has(identity)) {
-        repeated += 1;
-        const priorExit = rounds
-          .slice(0, round.index)
-          .reverse()
-          .find(({ reason }) => reason !== 'merged');
-        if (!seenInRound.has(identity) && priorExit) {
-          explained += 1;
-          explanations[priorExit.reason] =
-            (explanations[priorExit.reason] || 0) + 1;
-        } else {
-          unexplained += 1;
-        }
-      }
-      seen.add(identity);
-      seenInRound.add(identity);
-    }
-  }
-  return { repeated, explained, unexplained, explanations };
-}
-
 export function mergeQueueEvidence(events, runs, jobsByRun, mergedAt) {
   const orderedEvents = [...events]
     .filter(({ __typename }) =>
@@ -334,7 +304,7 @@ export function mergeQueueEvidence(events, runs, jobsByRun, mergedAt) {
       beforeCommit: event.beforeCommit?.oid || null,
       mergeGroupRuns: roundRuns.map((run) => ({
         id: run.id,
-        workflowIdentity: workflowIdentity(run),
+        workflowIdentity: queueEvidence.workflowIdentity(run),
         headSha: run.head_sha,
         createdAt: run.created_at,
         completedAt: run.updated_at,
@@ -448,7 +418,7 @@ export function mergeQueueEvidence(events, runs, jobsByRun, mergedAt) {
         exitedRounds.filter((round) => round.reason === reason).length,
       ]),
   );
-  const validationRepeats = validationRepeatEvidence(rounds);
+  const validationRepeats = queueEvidence.validationRepeatEvidence(rounds);
   return {
     queueStatus,
     status: deliveryComplete ? 'observed' : 'incomplete',
@@ -481,213 +451,6 @@ export function mergeQueueEvidence(events, runs, jobsByRun, mergedAt) {
     postDequeueRunnerMs: runnerEvidenceComplete ? postDequeueRunnerMs : null,
     rounds,
     diagnostics,
-  };
-}
-
-export function summarizeMergeQueueDelivery(samples) {
-  const queueObserved = samples.filter(
-    ({ mergeQueue }) => mergeQueue?.queueStatus === 'observed',
-  );
-  const deliveryObserved = queueObserved.filter(
-    ({ mergeQueue }) => mergeQueue?.status === 'observed',
-  );
-  const runnerObserved = queueObserved.filter(
-    ({ mergeQueue }) => mergeQueue.runnerEvidenceComplete,
-  );
-  const runnerWaitObserved = queueObserved.filter(
-    ({ mergeQueue }) => mergeQueue.runnerWaitEvidenceComplete,
-  );
-  const durations = deliveryObserved.map(
-    ({ mergeQueue }) => mergeQueue.deliveryDurationMs,
-  );
-  const dequeuePrCount = queueObserved.filter(
-    ({ mergeQueue }) => mergeQueue.dequeueCount > 0,
-  ).length;
-  const mergeConflictPrCount = queueObserved.filter(
-    ({ mergeQueue }) => (mergeQueue.dequeueReasons?.merge_conflict || 0) > 0,
-  ).length;
-  const reasons = {};
-  for (const { mergeQueue } of queueObserved) {
-    for (const [reason, count] of Object.entries(
-      mergeQueue.dequeueReasons || {},
-    )) {
-      reasons[reason] = (reasons[reason] || 0) + count;
-    }
-  }
-  const enoughSamples = deliveryObserved.length >= MINIMUM_SAMPLE_COUNT;
-  const p50Ms = nearestRank(durations, 0.5);
-  const p90Ms = nearestRank(durations, 0.9);
-  const dequeueRate = queueObserved.length
-    ? dequeuePrCount / queueObserved.length
-    : null;
-  const mergeConflictRate = queueObserved.length
-    ? mergeConflictPrCount / queueObserved.length
-    : null;
-  const unexplainedRepeatedValidationCount = queueObserved.reduce(
-    (total, { mergeQueue }) =>
-      total + (mergeQueue.unexplainedRepeatedValidationCount || 0),
-    0,
-  );
-  const runnerWaitUpperBounds = runnerWaitObserved.map(
-    ({ mergeQueue }) => mergeQueue.runnerWaitUpperBoundMs,
-  );
-  const runnerWaitQualified =
-    runnerWaitObserved.length === queueObserved.length &&
-    runnerWaitUpperBounds.length > 0 &&
-    Math.max(...runnerWaitUpperBounds) <= 30 * 60 * 1000;
-  const meetsTarget =
-    p50Ms !== null &&
-    p50Ms <= 15 * 60 * 1000 &&
-    p90Ms !== null &&
-    p90Ms <= 30 * 60 * 1000 &&
-    dequeueRate < 0.1 &&
-    mergeConflictRate < 0.05 &&
-    unexplainedRepeatedValidationCount === 0 &&
-    runnerWaitQualified;
-  return {
-    queueObservedCount: queueObserved.length,
-    deliveryObservedCount: deliveryObserved.length,
-    incompleteCount: samples.filter(
-      ({ mergeQueue }) => mergeQueue?.queueStatus === 'incomplete',
-    ).length,
-    notObservedCount: samples.filter(
-      ({ mergeQueue }) => mergeQueue?.queueStatus === 'not-observed',
-    ).length,
-    runnerEvidenceObservedCount: runnerObserved.length,
-    statistics: {
-      sampleCount: deliveryObserved.length,
-      p50Ms,
-      p90Ms,
-      maxMs: durations.length ? Math.max(...durations) : null,
-    },
-    dequeue: {
-      pullRequestCount: dequeuePrCount,
-      rate: dequeueRate,
-      exitCount: queueObserved.reduce(
-        (total, { mergeQueue }) => total + mergeQueue.dequeueCount,
-        0,
-      ),
-      reasons,
-      mergeConflict: {
-        pullRequestCount: mergeConflictPrCount,
-        rate: mergeConflictRate,
-        exclusiveMax: 0.05,
-      },
-    },
-    repeatedValidationCount: queueObserved.reduce(
-      (total, { mergeQueue }) => total + mergeQueue.repeatedValidationCount,
-      0,
-    ),
-    explainedRepeatedValidationCount: queueObserved.reduce(
-      (total, { mergeQueue }) =>
-        total + (mergeQueue.explainedRepeatedValidationCount || 0),
-      0,
-    ),
-    unexplainedRepeatedValidationCount,
-    runnerWait: {
-      authority:
-        'github-actions-workflow-created_at-to-job-started_at-upper-bound',
-      evidenceObservedCount: runnerWaitObserved.length,
-      statistics: {
-        sampleCount: runnerWaitUpperBounds.length,
-        p50Ms: nearestRank(runnerWaitUpperBounds, 0.5),
-        p95Ms: nearestRank(runnerWaitUpperBounds, 0.95),
-        maxMs: runnerWaitUpperBounds.length
-          ? Math.max(...runnerWaitUpperBounds)
-          : null,
-      },
-      target: { maxUpperBoundMs: 30 * 60 * 1000 },
-      qualified: runnerWaitQualified,
-    },
-    wastedRunnerMs: runnerObserved.reduce(
-      (total, { mergeQueue }) => total + mergeQueue.wastedRunnerMs,
-      0,
-    ),
-    postDequeueRunnerMs: runnerObserved.reduce(
-      (total, { mergeQueue }) => total + mergeQueue.postDequeueRunnerMs,
-      0,
-    ),
-    verdict: {
-      qualified: enoughSamples && meetsTarget,
-      reason: !deliveryObserved.length
-        ? 'no observed merge queue samples'
-        : !enoughSamples
-          ? 'insufficient merge queue sample count'
-          : !meetsTarget
-            ? 'merge queue delivery sample exceeds target'
-            : 'merge queue delivery sample meets target',
-    },
-  };
-}
-
-export function postMergeAdvisoryEvidence(pull, runs) {
-  const matching = runs
-    .filter(({ head_sha: headSha }) => headSha === pull.merge_commit_sha)
-    .sort(
-      (left, right) =>
-        new Date(left.created_at).getTime() -
-        new Date(right.created_at).getTime(),
-    );
-  const attempts = matching.map((run) => {
-    const terminal = run.status === 'completed' && Boolean(run.updated_at);
-    return {
-      workflowRunId: run.id,
-      status: terminal ? 'observed' : 'incomplete',
-      conclusion: run.conclusion || null,
-      createdAt: run.created_at || null,
-      completedAt: terminal ? run.updated_at : null,
-      durationMs:
-        terminal && run.created_at
-          ? Math.max(0, milliseconds(run.created_at, run.updated_at))
-          : null,
-      tailDurationMs:
-        terminal && pull.merged_at
-          ? Math.max(0, milliseconds(pull.merged_at, run.updated_at))
-          : null,
-    };
-  });
-  const complete = attempts.filter(({ status }) => status === 'observed');
-  return {
-    status: !attempts.length
-      ? 'not-observed'
-      : complete.length === attempts.length
-        ? 'observed'
-        : 'incomplete',
-    authority: 'github-actions-dev-post-merge-advisory-workflow-run',
-    criticalPathEligible: false,
-    mergeCriticalMetricImpact: 'excluded',
-    attemptCount: attempts.length,
-    tailDurationMs: complete.length
-      ? Math.max(...complete.map(({ tailDurationMs }) => tailDurationMs))
-      : null,
-    attempts,
-  };
-}
-
-export function summarizePostMergeAdvisory(samples) {
-  const evidence = samples
-    .map(({ postMergeAdvisory }) => postMergeAdvisory)
-    .filter(Boolean);
-  const observed = evidence.filter(({ status }) => status === 'observed');
-  const durations = observed
-    .map(({ tailDurationMs }) => tailDurationMs)
-    .filter(Number.isFinite);
-  return {
-    authority: 'github-actions-dev-post-merge-advisory-workflow-run',
-    criticalPathEligible: false,
-    mergeCriticalMetricImpact: 'excluded',
-    observedCount: observed.length,
-    incompleteCount: evidence.filter(({ status }) => status === 'incomplete')
-      .length,
-    notObservedCount: evidence.filter(({ status }) => status === 'not-observed')
-      .length,
-    statistics: {
-      sampleCount: durations.length,
-      p50Ms: nearestRank(durations, 0.5),
-      p95Ms: nearestRank(durations, 0.95),
-      maxMs: durations.length ? Math.max(...durations) : null,
-    },
-    samples: evidence,
   };
 }
 
@@ -2093,8 +1856,11 @@ export function report(
   cache.warmRatio = cacheObserved ? cache.warmCount / cacheObserved : null;
   cache.coldRatio = cacheObserved ? cache.coldCount / cacheObserved : null;
   const queueRecords = cohort.mergeQueueRecords;
-  const mergeQueueDelivery = summarizeMergeQueueDelivery(queueRecords);
-  const postMergeAdvisory = summarizePostMergeAdvisory(cohort.records);
+  const mergeQueueDelivery =
+    queueEvidence.summarizeMergeQueueDelivery(queueRecords);
+  const postMergeAdvisory = queueEvidence.summarizePostMergeAdvisory(
+    cohort.records,
+  );
   const deliveryEvidence = summarizeDeliveryEvidence(samples);
   const generatedAt = new Date().toISOString();
   const candidateTimelines = samples.map((sample) =>
@@ -2291,7 +2057,7 @@ async function main() {
         pull,
         requiredContexts,
         mergeQueueByPull.get(pull.number),
-        postMergeAdvisoryEvidence(pull, postMergeAdvisoryRuns),
+        queueEvidence.postMergeAdvisoryEvidence(pull, postMergeAdvisoryRuns),
         token,
         finalDevByPull.get(pull.number),
         options.latencyOnly,
